moved function to eda.py
